requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2-binary>=2.8.5,<2.9.0
djangorestframework-jwt>=1.11.0,<1.12.0